refactor: 훅과 서비스 파일들을 기능별 폴더로 구조 개선
- src/hooks/auth/ 폴더 생성하여 인증 관련 훅들 정리
  - useAuthForm.ts, useEmailVerification.ts, useProfileUpdate.ts 이동
- src/hooks/file-explorer/ 폴더 생성하여 파일 탐색기 관련 훅 정리
  - useFileTree.ts 이동
- src/services/ 폴더 생성하여 서비스 레이어 정리
  - file-system-service.ts → file-system.ts로 이동
